requirements.txt
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-lru==2.0.5
atomicwrites==1.4.0
contourpy==1.3.1
cycler==0.12.1
fonttools==4.57.0
fqdn==1.5.1
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imbalanced-learn==0.13.0
inflection==0.5.1
isoduration==20.11.0
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_server==2.27.3
kiwisolver==1.4.8
matplotlib==3.10.1
neurokit2==0.2.10
notebook==7.3.3
notebook_shim==0.2.4
numpy==2.2.4
overrides==7.7.0
pandas==2.2.3
pillow==11.1.0
ply==3.11
prometheus_client==0.21.1
pyls-spyder==0.4.0
pyparsing==3.2.3
# PyQt5 y PyQtWebEngine requieren entorno gráfico y dependencias pesadas.
# Si necesitas usarlos, instala con: pip install --only-binary :all: PyQt5==5.15.10
#PyQt5==5.15.10
#PyQtWebEngine==5.15.6
python-json-logger==3.3.0
pytz==2025.2
PyWavelets==1.8.0
pywin32==308; sys_platform == "win32"
pywinpty==2.0.15; sys_platform == "win32"
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.6.1
scipy==1.15.2
Send2Trash==1.8.3
sklearn-compat==0.1.3
sniffio==1.3.1
soundfile==0.13.1
terminado==0.18.1
threadpoolctl==3.6.0
types-python-dateutil==2.9.0.20241206
tzdata==2025.2
uri-template==1.3.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wfdb==4.3.0
xgboost==3.0.0
